feat: get top and bottom of the stack

+++ narsirabad.h
@@ -3,20 +3,21 @@
 #define NARSIRABAD
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <stdint.h>
 
 typedef struct Block {
     bool free;
     uint8_t offset;
-    uint32_t size;
+    size_t size;
     void* ptr;
 } Block;
 
 typedef struct Allocator {
     Block* headers;
     // These are counts of block, not amount of memory remaining
-    uint8_t header_len;
-    uint8_t header_capacity;
+    size_t header_len;
+    size_t header_capacity;
 } Allocator;
 
 void* allocate(uint32_t size);
